site/golf_federated/requirements.txt
Flask==1.1.2
Flask_SQLAlchemy==2.5.1
h5py==2.10.0
itsdangerous==1.1.0
loguru==0.5.3
numpy==1.22.3
pycryptodome==3.15.0
PyMySQL==1.0.2
PyYAML==6.0
redis==3.5.3
requests==2.24.0
scipy==1.7.3
sseclient==0.0.27
tensorflow==2.3.4
torch==1.8.1
